# first attempt at using SnoopCompile to speed things up