## current main

```bash
~/MixedModels.jl$ for i in {1..5}; do echo "Run $i"; julia --project=. -e'using MixedModels; @time fit(MixedModel, @formula(reaction ~ 1 + days + (1+days|subj)), MixedModels.dataset(:sleepstudy)); @time fit(MixedModel,@formula(use ~ 1+age+abs2(age)+urban+livch+(1|urban&dist)), MixedModels.dataset(:contra), Bernoulli(), LogitLink())'; done
Run 1
[ Info: using MKL.jl
Minimizing 58    Time: 0:00:00 (13.88 ms/it)
 22.300008 seconds (77.44 M allocations: 4.614 GiB, 6.05% gc time, 99.83% compilation time)
Minimizing 254   Time: 0:00:00 ( 1.10 ms/it)
 16.647709 seconds (49.34 M allocations: 2.975 GiB, 4.31% gc time, 99.16% compilation time)
Run 2
[ Info: using MKL.jl
Minimizing 58    Time: 0:00:00 (12.15 ms/it)
 22.237959 seconds (77.44 M allocations: 4.614 GiB, 5.52% gc time, 99.84% compilation time)
Minimizing 254   Time: 0:00:00 ( 1.27 ms/it)
 16.607973 seconds (49.34 M allocations: 2.975 GiB, 4.44% gc time, 98.91% compilation time)
Run 3
[ Info: using MKL.jl
Minimizing 58    Time: 0:00:00 (12.33 ms/it)
 21.382303 seconds (77.44 M allocations: 4.614 GiB, 5.54% gc time, 99.84% compilation time)
Minimizing 254   Time: 0:00:00 ( 1.09 ms/it)
 16.072365 seconds (49.34 M allocations: 2.975 GiB, 4.35% gc time, 99.13% compilation time)
Run 4
[ Info: using MKL.jl
Minimizing 58    Time: 0:00:00 (12.55 ms/it)
 21.785942 seconds (77.44 M allocations: 4.614 GiB, 5.47% gc time, 99.83% compilation time)
Minimizing 254   Time: 0:00:00 ( 1.12 ms/it)
 16.510533 seconds (49.34 M allocations: 2.975 GiB, 4.26% gc time, 99.15% compilation time)
Run 5
[ Info: using MKL.jl
Minimizing 58    Time: 0:00:00 (12.30 ms/it)
 22.517317 seconds (77.44 M allocations: 4.614 GiB, 5.32% gc time, 99.84% compilation time)
Minimizing 254   Time: 0:00:00 ( 1.19 ms/it)
 16.357725 seconds (49.34 M allocations: 2.975 GiB, 4.41% gc time, 99.14% compilation time)
```

## with new precompile
```bash
~/MixedModels.jl$ for i in {1..5}; do echo "Run $i"; julia --project=. -e'using MixedModels; @time fit(MixedModel, @formula(reaction ~ 1 + days + (1+days|subj)), MixedModels.dataset(:sleepstudy)); @time fit(MixedModel,@formula(use ~ 1+age+abs2(age)+urban+livch+(1|urban&dist)), MixedModels.dataset(:contra), Bernoulli(), LogitLink())'; done
[ Info: using MKL.jl
Minimizing 58    Time: 0:00:00 ( 6.42 ms/it)
 13.022910 seconds (42.46 M allocations: 2.437 GiB, 4.60% gc time, 99.66% compilation time)
Minimizing 254   Time: 0:00:00 ( 1.15 ms/it)
 12.423930 seconds (38.11 M allocations: 2.248 GiB, 5.15% gc time, 98.81% compilation time)
Run 2
[ Info: using MKL.jl
Minimizing 58    Time: 0:00:00 ( 6.04 ms/it)
 12.812662 seconds (42.46 M allocations: 2.437 GiB, 4.50% gc time, 99.72% compilation time)
Minimizing 254   Time: 0:00:00 ( 1.20 ms/it)
 12.360282 seconds (38.11 M allocations: 2.248 GiB, 5.20% gc time, 98.79% compilation time)
Run 3
[ Info: using MKL.jl
Minimizing 58    Time: 0:00:00 ( 6.42 ms/it)
 14.338150 seconds (42.46 M allocations: 2.437 GiB, 4.70% gc time, 99.75% compilation time)
Minimizing 254   Time: 0:00:00 ( 1.19 ms/it)
 13.086941 seconds (38.11 M allocations: 2.248 GiB, 5.23% gc time, 98.88% compilation time)
Run 4
[ Info: using MKL.jl
Minimizing 58    Time: 0:00:00 ( 6.24 ms/it)
 13.455821 seconds (42.46 M allocations: 2.437 GiB, 4.52% gc time, 99.72% compilation time)
Minimizing 254   Time: 0:00:00 ( 1.53 ms/it)
 14.986385 seconds (38.11 M allocations: 2.248 GiB, 5.07% gc time, 98.75% compilation time)
Run 5
[ Info: using MKL.jl
Minimizing 58    Time: 0:00:00 ( 6.35 ms/it)
 12.882482 seconds (42.46 M allocations: 2.437 GiB, 4.56% gc time, 99.71% compilation time)
Minimizing 254   Time: 0:00:00 ( 1.16 ms/it)
 12.763560 seconds (38.12 M allocations: 2.248 GiB, 5.20% gc time, 98.88% compilation time)
```

## how I did this
```julia
using SnoopCompile, MixedModels
ttot, pcs = SnoopCompile.parcel(snoop)
snoop = @snoopi_deep  begin 
       MixedModels.fit(MixedModel, @formula(reaction ~ 1 + days + (1+days|subj)), MixedModels.dataset(:sleepstudy))
       MixedModels.fit(MixedModel,@formula(use ~ 1+age+abs2(age)+urban+livch+(1|urban&dist)), MixedModels.dataset(:contra), Bernoulli())
       end
SnoopCompile.write(joinpath(@__DIR__, "src"), pcs;has_bodyfunction=true)
```

I then had to edit `precompile_MixedModels.jl` and replace a few type parameters that were still there (`s/T<:Core.AbstractFloat/Float64/g` and `s/D<:Distributions.Distribution/Distributions.Bernoulli/g`).

Because of the way Tables.jl and StatsModels use NamedTuples, I also had to remove lines referencing specific variables in these models.

## PR template
Did behavior change? Did you add need features? If so, please update NEWS.md
- [x] add entry in NEWS.md
- [x] after opening this PR, add a reference and run `docs/NEWS-update.jl` to update the cross-references.

Should we release your changes right away? If so, bump the version:
- [x] I've bumped the version appropriately