@dmbates 2022-05-01
> @dmbates should we turn off autoformatting for the precompile files? They're generated by SnoopCompile and if we leave them as emitted then it makes it easier to compare things when we do more inspection with Snoop.

Agreed.