@palday 2022-05-02
Here some final timings using `benchmarks/snoopspeed.sh` (which uses the kb07 and verbagg datasets, so is more of a test of generalization than the previous tests):

## current main
```bash
~/MixedModels.jl$ time benchmark/snoopspeed.sh
Run 1
[ Info: using MKL.jl
Minimizing 93    Time: 0:00:01 (13.39 ms/it)
 25.721185 seconds (88.90 M allocations: 5.222 GiB, 5.69% gc time, 99.73% compilation time)
[ Info: using MKL.jl
Minimizing 203   Time: 0:00:01 ( 5.06 ms/it)
 34.859563 seconds (117.98 M allocations: 7.119 GiB, 5.12% gc time, 97.94% compilation time)
Run 2
[ Info: using MKL.jl
Minimizing 93    Time: 0:00:01 (12.92 ms/it)
 25.323026 seconds (88.90 M allocations: 5.222 GiB, 5.24% gc time, 99.77% compilation time)
[ Info: using MKL.jl
Minimizing 203   Time: 0:00:01 ( 5.09 ms/it)
 35.383858 seconds (117.98 M allocations: 7.119 GiB, 5.17% gc time, 97.96% compilation time)
Run 3
[ Info: using MKL.jl
Minimizing 93    Time: 0:00:01 (16.60 ms/it)
 26.033828 seconds (88.90 M allocations: 5.222 GiB, 5.14% gc time, 99.77% compilation time)
[ Info: using MKL.jl
Minimizing 203   Time: 0:00:01 ( 5.02 ms/it)
 37.257851 seconds (117.98 M allocations: 7.119 GiB, 5.12% gc time, 98.06% compilation time)

real    3m40.261s
user    4m2.498s
sys     0m7.523s
```

## this PR
```bash
~/MixedModels.jl$ time benchmark/snoopspeed.sh
Run 1
[ Info: using MKL.jl
Minimizing 93    Time: 0:00:01 (12.58 ms/it)
 19.097108 seconds (55.30 M allocations: 3.117 GiB, 3.88% gc time, 99.67% compilation time)
[ Info: using MKL.jl
Minimizing 203   Time: 0:00:01 ( 5.41 ms/it)
 25.843597 seconds (75.05 M allocations: 4.380 GiB, 4.62% gc time, 97.07% compilation time)
Run 2
[ Info: using MKL.jl
Minimizing 93    Time: 0:00:00 ( 9.59 ms/it)
 17.110819 seconds (55.30 M allocations: 3.117 GiB, 4.38% gc time, 99.65% compilation time)
[ Info: using MKL.jl
Minimizing 203   Time: 0:00:01 ( 5.52 ms/it)
 23.983724 seconds (75.05 M allocations: 4.380 GiB, 4.67% gc time, 96.65% compilation time)
Run 3
[ Info: using MKL.jl
Minimizing 93    Time: 0:00:00 ( 9.28 ms/it)
 16.480201 seconds (55.30 M allocations: 3.117 GiB, 4.36% gc time, 99.66% compilation time)
[ Info: using MKL.jl
Minimizing 203   Time: 0:00:01 ( 5.29 ms/it)
 24.536892 seconds (75.05 M allocations: 4.380 GiB, 4.80% gc time, 96.97% compilation time)

real    3m4.099s
user    3m27.271s
sys     0m6.946s
```